requirements/requirements.txt
fastapi~=0.104.1
pydantic~=2.5.2
pydantic-settings~=2.1.0
SQLAlchemy~=2.0.23
psycopg2-binary~=2.9.9
asyncpg~=0.29.0
alembic~=1.12.1
alembic-postgresql-enum~=0.4.0
typer~=0.9.0
uvicorn~=0.24.0.post1
json_logging~=1.3.0
gunicorn~=21.2.0
starlette_context~=0.3.6
bcrypt~=4.1.2
python-jose~=3.3.0